Fixes some bad XML tags

+++ WPILib/Utility.cs
@@ -34,10 +34,10 @@ namespace WPILib
         /// <summary>
         /// Read the microsecond timer from the FPGA
         /// </summary>
-        public static long GetFPGATime()
+        public static uint GetFPGATime()
         {
             int status = 0;
-            long value = HAL.Base.HAL.GetFPGATime(ref status);
+            uint value = HAL.Base.HAL.GetFPGATime(ref status);
             return value;
         }
